feat(typescript): finish up typescript support
