+++ .dockerignore
@@ -6,3 +6,4 @@ node_modules/
 template/build
 template/coverage
 template/dist
+build/
