+++ .eslintignore
@@ -2,3 +2,4 @@ lib/formatters/**/*.js
 template/**/*.js
 node_modules/**/*.js
 tests/sink/**/*.js
+build/**/*.js
