+++ .gitignore
@@ -3,3 +3,4 @@ yarn-error.log
 .vscode/
 package-lock.json
 .DS_Store
+build/
