--- lib/jest/babelPreprocessor.js
@@ -1,15 +1,0 @@
-const babelJest = require('babel-jest')
-
-module.exports = babelJest.createTransformer({
-    presets: [
-        [
-            '@babel/preset-env',
-            {
-                targets: {
-                    node: 'current',
-                },
-            },
-        ],
-    ],
-    babelrc: false,
-})
