--- lib/util/platform.js
@@ -1,7 +1,0 @@
-const os = require('os')
-
-const isWindows = os.platform() === 'win32'
-
-module.exports = {
-    isWindows,
-}
