+++ package.json
@@ -1,9 +1,11 @@
 {
   "name": "typescript-node-scripts",
-  "version": "1.3.0",
+  "version": "2.0.0-beta.0",
   "description": "Create Node.js apps based on TypeScript with zero-configuration.",
   "main": "bin/typescript-node-scripts.js",
   "scripts": {
+    "watch": "tsc -p . --watch",
+    "compile": "tsc -p .",
     "test": "node bin/typescript-node-scripts.js test",
     "start": "node bin/typescript-node-scripts.js start",
     "build": "node bin/typescript-node-scripts.js build",
@@ -12,7 +14,8 @@
     "e2e:local": "docker build -t tns-docker . && docker run tns-docker",
     "lint": "eslint .",
     "commit": "git-cz",
-    "version": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md"
+    "version": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
+    "prepublishOnly": "yarn compile"
   },
   "bin": {
     "typescript-node-scripts": "./bin/typescript-node-scripts.js"
@@ -30,8 +33,8 @@
     "babel-loader": "8.0.5",
     "chalk": "2.4.2",
     "filesize": "4.1.2",
-    "fs-extra": "7.0.1",
     "fork-ts-checker-webpack-plugin": "1.0.0",
+    "fs-extra": "7.0.1",
     "jest": "24.5.0",
     "pstree.remy": "1.1.2",
     "strip-ansi": "5.0.0",
@@ -49,10 +52,13 @@
     "@commitlint/cli": "^7.2.1",
     "@commitlint/config-conventional": "^7.1.2",
     "@types/babel-code-frame": "^6.20.2",
+    "@types/filesize": "^4.1.0",
+    "@types/fs-extra": "^5.0.5",
     "@types/jest": "^23.3.10",
     "@types/node": "^10.12.11",
     "@types/text-table": "^0.2.0",
     "@types/webpack": "^4.4.24",
+    "@types/webpack-node-externals": "^1.6.3",
     "commitizen": "^3.0.7",
     "conventional-changelog-cli": "^2.0.12",
     "cz-conventional-changelog": "^2.1.0",
