+++ src/injectors/dev.ts
@@ -1,0 +1,5 @@
+process.env.NODE_ENV = 'development'
+
+process.on('unhandledRejection', err => {
+  throw err
+})
