+++ src/injectors/test.ts
@@ -1,0 +1,4 @@
+process.env.NODE_ENV = 'test'
+process.on('unhandledRejection', err => {
+  throw err
+})
