+++ src/lib/clearConsole.ts
