+++ src/lib/formatters/lintTableFormatter.ts
