+++ src/lib/jest/babelPreprocessor.ts
@@ -1,0 +1,15 @@
+const babelJest = require('babel-jest')
+
+module.exports = babelJest.createTransformer({
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        targets: {
+          node: 'current',
+        },
+      },
+    ],
+  ],
+  babelrc: false,
+})
