+++ src/lib/tsLintHelper.ts
