+++ src/types/TNS.ts
@@ -1,0 +1,6 @@
+export interface TNSOptions {
+  noCollapse?: boolean
+  noAutoStart?: boolean
+  useMonorepo?: boolean
+  bypassCiWarnings?: boolean
+}
